haircut-classifier/requirements.txt
# Core ML
torch>=2.2
torchvision>=0.17
open_clip_torch>=2.24
# MobileCLIP for later deploy export (pip package name per Apple repo)
# Install from: pip install git+https://github.com/apple/ml-mobileclip.git
# Listed here as a comment because it is not on PyPI as of writing.

# Data / training
pandas>=2.0
numpy>=1.24
pillow>=10.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66

# Preprocessing
mediapipe>=0.10
opencv-python>=4.9

# Serving
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9

# Tracking (optional, gated in code)
wandb>=0.16

# Tests
pytest>=8.0
